requirements.txt
gpt-2-simple==0.7.1
tensorflow==1.14
python-Levenshtein==0.12.0